refactor(templates): split template types/paginator into dedicated modules
Address latest review:
- JS: move TemplateInfo/TemplateListOpts into template/types.ts and
  TemplatePaginator into template/paginator.ts; delete templateApi.ts.
- Python: move TemplateInfo into template/types.py; delete template_api.py.
- CLI: simplify toTemplateSchema (spread, convert only date fields) and drop
  the 'Showing first N' hint (keep --limit).
