+++ packages/cli/src/commands/template/publish.ts
@@ -60,7 +60,7 @@ async function templateAction(
         template_id: template,
       })
     } else if (opts.select) {
-      const { templates: allTemplates } = await listSandboxTemplates()
+      const allTemplates = await listSandboxTemplates()
 
       const filteredTemplates = allTemplates.filter(
         (e) => !e.public === publish
